+++ src/ssh.rs
@@ -3,9 +3,13 @@ use crate::util;
 
 use anyhow::{bail, Context, Result};
 use indicatif::ProgressBar;
+use itertools::Itertools;
 use log::{debug, error, info};
+use regex::Regex;
 use rpassword::prompt_password_stderr;
+use ssh2::FileStat;
 use ssh2::Session as RawSession;
+use std::collections::HashMap;
 use std::fs::File;
 use std::io::prelude::*;
 use std::io::{BufReader, Error as IOError, ErrorKind};
@@ -27,6 +31,28 @@ pub struct SshSession<'a> {
 }
 
 impl<'a> SshSession<'a> {
+    /// Adjust the group of the remote target recursively
+    pub fn adjust_group(&self, file: &Path, group: &str) -> Result<()> {
+        let cmd = format!("chown -R :{} \"{}\"", group, file.display());
+        let mut channel = self.raw.channel_session()?;
+
+        channel.exec(&cmd).context("Could not adjust group.")?;
+        Ok(())
+    }
+
+    /// List all files present (relative to the current host's base-folder).
+    pub fn all_files(&self) -> Result<Vec<PathBuf>> {
+        let cmd = format!("cd {} && ls -1rt */*", self.host.folder.display());
+        let mut channel = self.raw.channel_session()?;
+
+        channel
+            .exec(&cmd)
+            .context("Could not execute listing command")?;
+        let mut files = String::new();
+        channel.read_to_string(&mut files)?;
+        Ok(files.lines().map(|s| Path::new(s).to_path_buf()).collect())
+    }
+
     fn auth(&self, auth: &Auth) -> Result<()> {
         if auth.use_agent {
             self.auth_agent()?;
@@ -119,15 +145,6 @@ impl<'a> SshSession<'a> {
         Ok(())
     }
 
-    /// Adjust the group of the remote target recursively
-    pub fn adjust_group(&self, file: &Path, group: &str) -> Result<()> {
-        let cmd = format!("chown -R :{} \"{}\"", group, file.display());
-        let mut channel = self.raw.channel_session()?;
-
-        channel.exec(&cmd).context("Could not adjust group.")?;
-        Ok(())
-    }
-
     /// Create a new SSH session from the given host configuration.
     ///
     /// First try authenticating with all agent identities then use an interactive password, if enabled.
@@ -155,58 +172,9 @@ impl<'a> SshSession<'a> {
         }
     }
 
-    /// List all files present (relative to the current host's base-folder).
-    pub fn list_files(&self) -> Result<Vec<PathBuf>> {
-        let cmd = format!("cd {} && ls -1rt */*", self.host.folder.display());
-        let mut channel = self.raw.channel_session()?;
-
-        channel
-            .exec(&cmd)
-            .context("Could not execute listing command")?;
-        let mut files = String::new();
-        channel.read_to_string(&mut files)?;
-        Ok(files.lines().map(|s| Path::new(s).to_path_buf()).collect())
-    }
-
-    pub fn get_files_by(
-        &self,
-        indices: &[i64],
-        names: &[&str],
-        prefix_length: u8,
-    ) -> Result<FileListing> {
-        let remote_files: Vec<(usize, PathBuf)> =
-            self.list_files()?.into_iter().enumerate().collect();
-        let num_files = remote_files.len() as i64;
-
-        let mut selected: Vec<(usize, PathBuf)> = Vec::new();
-
-        if indices.len() == 0 && names.len() == 0 {
-            for idx in 0..num_files {
-                selected.push(remote_files[idx as usize].clone());
-            }
-        }
-
-        for idx in indices.iter() {
-            let idx = if *idx < 0 { num_files + *idx } else { *idx } as usize;
-            selected.push(remote_files[idx as usize].clone());
-        }
-
-        'outer: for file in names {
-            let hash = util::get_hash(Path::new(file), prefix_length)?;
-            for file in remote_files.iter() {
-                if file.1.starts_with(&hash) {
-                    selected.push(file.clone());
-                    continue 'outer;
-                }
-            }
-            bail!("No file with same hash found on server: {}", file);
-        }
-
-        selected.sort_by_key(|e| e.0);
-        Ok(FileListing {
-            num_files: num_files as usize,
-            files: selected,
-        })
+    /// Get listing of files
+    pub fn list_files(&self) -> Result<FileListing> {
+        FileListing::new(&self)
     }
 
     /// Make folder on the remote site if it does not exist (relative to the current host's
@@ -286,10 +254,7 @@ impl<'a> SshSession<'a> {
     }
 
     /// Get stats about a remote files (relative to the current host's base-folder)
-    pub fn stat<'b, I: IntoIterator<Item = &'b Path>>(
-        &self,
-        paths: I,
-    ) -> Result<Vec<ssh2::FileStat>> {
+    pub fn stat<'b, I: IntoIterator<Item = &'b Path>>(&self, paths: I) -> Result<Vec<FileStat>> {
         debug!("Getting remote stats…");
         let paths: Vec<_> = paths.into_iter().collect();
 
@@ -350,7 +315,217 @@ impl<'a> SshSession<'a> {
     }
 }
 
-pub struct FileListing {
+/// Helper structure to avoid re-implementing file listing capabilities for all commands.
+pub struct FileListing<'a> {
     pub num_files: usize,
-    pub files: Vec<(usize, PathBuf)>,
+    all_files: HashMap<usize, PathBuf>,
+    pub indices: Vec<usize>,
+    pub stats: Option<HashMap<usize, FileStat>>,
+    ssh: &'a SshSession<'a>,
+}
+
+impl<'a> FileListing<'a> {
+    fn new(ssh: &'a SshSession) -> Result<FileListing<'a>> {
+        let all_files: HashMap<_, _> = ssh.all_files()?.into_iter().enumerate().collect();
+        let num_files = all_files.len();
+
+        Ok(Self {
+            num_files,
+            all_files,
+            indices: Vec::new(),
+            stats: None,
+            ssh,
+        })
+    }
+
+    /// Simply select all files if argument is true
+    pub fn add_all(mut self, select_all: bool) -> Self {
+        if select_all {
+            let mut all: Vec<usize> = (0..self.num_files).collect();
+            self.indices.append(&mut all);
+            self.indices = Self::make_unique(self.indices.drain(..));
+        }
+        self
+    }
+
+    /// Select all files which name matches regex
+    pub fn by_filter(self, filter: Option<&str>) -> Result<Self> {
+        match filter {
+            None => Ok(self),
+            Some(filter) => {
+                let re = Regex::new(filter)?;
+                let indices = {
+                    let mut indices = self.indices;
+                    let mut additions: Vec<_> = self
+                        .all_files
+                        .iter()
+                        .map(|(idx, p)| (idx, p.as_path()))
+                        .filter(|(_, path)| {
+                            re.is_match(&path.file_name().unwrap().to_string_lossy().to_string())
+                        })
+                        .map(|(idx, _)| *idx)
+                        .collect();
+
+                    indices.append(&mut additions);
+                    Self::make_unique(indices)
+                };
+                Ok(Self { indices, ..self })
+            }
+        }
+    }
+
+    /// Select all files with corresponding indices
+    pub fn by_indices(self, indices: &[i64]) -> Result<Self> {
+        if indices.len() > 0 {
+            let num_files = self.num_files as i64;
+            for idx in indices {
+                if *idx < -num_files || *idx >= num_files {
+                    bail!("Invalid index specified: {}", idx);
+                }
+            }
+
+            let num_files = self.num_files as i64;
+
+            let indices = {
+                let mut self_indices = self.indices;
+                let mut additions: Vec<_> = indices
+                    .iter()
+                    .map(|idx| if *idx < 0 { num_files + *idx } else { *idx } as usize)
+                    .collect();
+                self_indices.append(&mut additions);
+                Self::make_unique(self_indices)
+            };
+            Ok(Self { indices, ..self })
+        } else {
+            Ok(self)
+        }
+    }
+
+    /// Select all files that have the same hash as the names given
+    pub fn by_name(self, names: &[&str], prefix_length: u8) -> Result<Self> {
+        if names.len() > 0 {
+            let indices = {
+                let mut indices = self.indices;
+
+                let hash_to_file: HashMap<String, usize> = self
+                    .all_files
+                    .iter()
+                    .filter(|(_, path)| path.parent().is_some())
+                    .map(|(idx, path)| {
+                        let prefix = path.parent().unwrap();
+                        let truncated_prefix = prefix
+                            .to_string_lossy()
+                            .chars()
+                            .take(prefix_length as usize)
+                            .collect();
+                        (truncated_prefix, *idx)
+                    })
+                    .collect();
+
+                for file in names {
+                    let hash = util::get_hash(Path::new(file), prefix_length)?;
+                    match hash_to_file.get(&hash) {
+                        Some(idx) => indices.push(*idx),
+                        None => bail!("No file with same hash found on server: {}", file),
+                    }
+                }
+                Self::make_unique(indices)
+            };
+            Ok(Self { indices, ..self })
+        } else {
+            Ok(self)
+        }
+    }
+
+    pub fn iter(&'a self) -> Result<FileListingIter<'a>> {
+        let stats = self.stats.as_ref();
+        let paths = &self.all_files;
+        Ok(FileListingIter::new(&self.indices[..], paths, stats))
+    }
+
+    /// Only use last `n` files
+    pub fn last(self, n: Option<usize>) -> Self {
+        match n {
+            Some(n) => {
+                let indices = self.indices.into_iter().skip(self.num_files - n).collect();
+                Self { indices, ..self }
+            }
+            None => self,
+        }
+    }
+
+    pub fn revert(mut self, do_revert: bool) -> Self {
+        if do_revert {
+            self.indices.reverse();
+        }
+        self
+    }
+
+    pub fn sort_by_size(mut self, sort_by_size: bool) -> Result<Self> {
+        if sort_by_size {
+            self.ensure_stats()?;
+            let stats = self.stats.as_ref().unwrap();
+            self.indices
+                .sort_by_key(|idx| stats.get(idx).unwrap().size.unwrap());
+        }
+        Ok(self)
+    }
+
+    pub fn with_stats(mut self, with_stats: bool) -> Result<Self> {
+        if with_stats {
+            self.ensure_stats()?;
+        }
+        Ok(self)
+    }
+
+    fn ensure_stats(&mut self) -> Result<()> {
+        if self.stats.is_none() {
+            let paths = self
+                .indices
+                .iter()
+                .map(|i| self.all_files.get(i).unwrap().as_path());
+            let idx = self.indices.iter().map(|idx| *idx);
+            let raw_stats = self.ssh.stat(paths)?;
+            self.stats = Some(idx.zip(raw_stats.into_iter()).collect());
+        }
+        Ok(())
+    }
+
+    fn make_unique<I: IntoIterator<Item = usize>>(indices: I) -> Vec<usize> {
+        indices.into_iter().unique().collect()
+    }
+}
+
+pub struct FileListingIter<'a> {
+    iter_idx: std::slice::Iter<'a, usize>,
+    files: &'a HashMap<usize, PathBuf>,
+    stats: Option<&'a HashMap<usize, FileStat>>,
+}
+
+impl<'a> FileListingIter<'a> {
+    fn new(
+        indices: &'a [usize],
+        files: &'a HashMap<usize, PathBuf>,
+        stats: Option<&'a HashMap<usize, FileStat>>,
+    ) -> Self {
+        Self {
+            iter_idx: indices.iter(),
+            files,
+            stats,
+        }
+    }
+}
+
+impl<'a> Iterator for FileListingIter<'a> {
+    type Item = (usize, &'a Path, Option<&'a FileStat>);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let (idx, file) = match self.iter_idx.next().cloned() {
+            Some(idx) => (idx, self.files.get(&idx).unwrap()),
+            None => return None,
+        };
+        let stat = self.stats.map(|s| s.get(&idx).unwrap());
+
+        Some((idx, file, stat))
+    }
 }